requirements.txt
matplotlib==1.5.3
pandas==0.20.3
numpy==1.13.3
scipy==0.18.1
tensorflow_gpu==1.5.0
scikit_learn==0.19.1
jupyter==5.4.0
